# swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Your API Title",
      "description": "Description of your API",
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:3000",
        "description": "Development server"
      }
    ],
    "paths": {
...
      "/auth/refresh-token": {
        "post": {
          "summary": "Refresh JWT token",
          "operationId": "refreshToken",
          "tags": ["Authentication"],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Token refreshed successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "token": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
...
      "/file/upload": {
        "post": {
          "summary": "Upload file",
          "operationId": "uploadFile",
          "tags": ["File"],
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "file": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "File uploaded successfully"
            },
            "400": {
              "description": "Bad request"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/file/download/{filename}": {
        "get": {
          "summary": "Download file",
          "operationId": "downloadFile",
          "tags": ["File"],
          "parameters": [
            {
              "name": "filename",
              "in": "path",
              "description": "Filename",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "File download successful"
            },
            "404": {
              "description": "File not found"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
...
        "RefreshTokenRequest": {
          "type": "object",
          "properties": {
            "refreshToken": {
              "type": "string"
            }
          }
        },
...
      }
    }
  }